requirements.txt
gymnasium
playwright
Pillow
evaluate
openai==0.27.0
types-tqdm
tiktoken
aiolimiter
beartype
flask
nltk
text-generation
transformers==4.33.2
python-dotenv
tqdm